refactor: application 관련 Type 클래스 패키지 이동
- VerifyStatus 을 application.domain 로
