+++ src/main/java/com/example/solidconnection/admin/dto/GpaScoreResponse.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.admin.dto;
 
 import com.example.solidconnection.score.domain.GpaScore;
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 
 public record GpaScoreResponse(
         long id,
