+++ src/main/java/com/example/solidconnection/admin/dto/GpaScoreUpdateRequest.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.admin.dto;
 
 import com.example.solidconnection.custom.validation.annotation.RejectedReasonRequired;
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 import jakarta.validation.constraints.NotNull;
 
 @RejectedReasonRequired
