+++ src/main/java/com/example/solidconnection/admin/dto/LanguageTestScoreResponse.java
@@ -2,7 +2,7 @@ package com.example.solidconnection.admin.dto;
 
 import com.example.solidconnection.score.domain.LanguageTestScore;
 import com.example.solidconnection.university.domain.LanguageTestType;
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 
 public record LanguageTestScoreResponse(
         long id,
