+++ src/main/java/com/example/solidconnection/admin/dto/LanguageTestScoreUpdateRequest.java
@@ -2,7 +2,7 @@ package com.example.solidconnection.admin.dto;
 
 import com.example.solidconnection.custom.validation.annotation.RejectedReasonRequired;
 import com.example.solidconnection.university.domain.LanguageTestType;
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 import jakarta.validation.constraints.NotNull;
 
 @RejectedReasonRequired
