+++ src/main/java/com/example/solidconnection/admin/dto/ScoreSearchCondition.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.admin.dto;
 
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 
 import java.time.LocalDate;
