+++ src/main/java/com/example/solidconnection/application/domain/VerifyStatus.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.type;
+package com.example.solidconnection.application.domain;
 
 public enum VerifyStatus {
     PENDING, REJECTED, APPROVED
