+++ src/main/java/com/example/solidconnection/score/domain/GpaScore.java
@@ -3,7 +3,7 @@ package com.example.solidconnection.score.domain;
 import com.example.solidconnection.application.domain.Gpa;
 import com.example.solidconnection.entity.common.BaseEntity;
 import com.example.solidconnection.siteuser.domain.SiteUser;
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
