+++ src/main/java/com/example/solidconnection/score/dto/GpaScoreStatusResponse.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.score.dto;
 
 import com.example.solidconnection.score.domain.GpaScore;
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 
 public record GpaScoreStatusResponse(
         long id,
