+++ src/test/java/com/example/solidconnection/application/service/ApplicationQueryServiceTest.java
@@ -14,7 +14,7 @@ import com.example.solidconnection.score.repository.LanguageTestScoreRepository;
 import com.example.solidconnection.siteuser.domain.SiteUser;
 import com.example.solidconnection.support.integration.BaseIntegrationTest;
 import com.example.solidconnection.university.domain.LanguageTestType;
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 import com.example.solidconnection.university.domain.UniversityInfoForApply;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
