+++ src/test/java/com/example/solidconnection/custom/validation/validator/RejectedReasonValidatorTest.java
@@ -3,7 +3,7 @@ package com.example.solidconnection.custom.validation.validator;
 import com.example.solidconnection.admin.dto.GpaScoreUpdateRequest;
 import com.example.solidconnection.admin.dto.LanguageTestScoreUpdateRequest;
 import com.example.solidconnection.university.domain.LanguageTestType;
-import com.example.solidconnection.type.VerifyStatus;
+import com.example.solidconnection.application.domain.VerifyStatus;
 import jakarta.validation.ConstraintViolation;
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
